# public/openapi.yaml
openapi: 3.0.3
info:
  title: 同频 B2 API
  version: 1.3.0
  description: 同频 B2 情绪急救与数字陪伴前后端接口。支持 Swagger UI Try it out；可在 Servers 中切换本地或远程环境。注意：本 API 不替代专业医疗或心理服务，AI 输出必须人工确认。
servers:
- url: http://localhost:8080/api/v1
  description: 本地开发环境
- url: https://{remoteHost}/api/v1
  description: 远程环境（修改 remoteHost 为实际域名）
  variables:
    remoteHost:
      default: api.tongpin.example.com
      description: 远程 API 域名，不带 https:// 和路径
...
paths:
...
  /me:
    get:
      tags:
      - 用户端
      summary: 获取当前用户信息
      operationId: getCurrentUser
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
      - 用户端
      summary: 更新昵称、年龄段和偏好
      operationId: updateCurrentUser
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              display_name: 小屿
              age_band: 14-17
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/consents:
    get:
      tags:
      - 用户端
      summary: 获取授权状态
      operationId: getConsents
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consents'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
      - 用户端
      summary: 更新设备、提醒、摘要、日记授权
      operationId: updateConsents
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              device_metrics: true
              activity_reminder: true
              trend_summary_share: false
              diary_share: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consents'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /me/device/trends:
    get:
      tags:
      - 用户端
      summary: 获取近 7 天设备趋势
      operationId: getDeviceTrends
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/device/care-responses:
    post:
      tags:
      - 用户端
      summary: 回应活动关怀提醒
      operationId: createDeviceCareResponse
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              scenario: body_uncomfortable
              want_exercise: false
              note: 今天身体不舒服
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/diary:
    get:
      tags:
      - 用户端
      summary: 获取日记列表
      operationId: listDiary
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/diary/drafts:
    post:
      tags:
      - 用户端
      summary: 生成或创建日记草稿
      operationId: createDiaryDraft
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              feeling: low
              free_text: 今天有点累，也不想说话。
              source: user_input
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiaryEntry'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/diary/{diary_id}:
    patch:
      tags:
      - 用户端
      summary: 编辑日记
      operationId: updateDiary
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - name: diary_id
        in: path
        required: true
        schema:
          type: string
        description: diary_id 标识
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              content: 今天有点累，但我想记录下来。
              feeling: low
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiaryEntry'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/diary/{diary_id}/confirm:
    post:
      tags:
      - 用户端
      summary: 确认日记
      operationId: confirmDiary
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - name: diary_id
        in: path
        required: true
        schema:
          type: string
        description: diary_id 标识
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              confirmed: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiaryEntry'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入登录后获取的 access token。匿名用户先调用 /auth/user-session。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: 写接口幂等键。重复提交同一请求时应返回同一结果。
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    User:
      type: object
      required:
      - id
      - age_band
      - district_id
      properties:
        id:
          type: string
        display_name:
          type: string
        age_band:
          type: string
        district_id:
          type: string
        created_at:
          type: string
          format: date-time
        consents:
          type: object
          additionalProperties: true
    Consents:
      type: object
      properties:
        device_metrics:
          type: boolean
        activity_reminder:
          type: boolean
        trend_summary_share:
          type: boolean
        diary_share:
          type: boolean
...
    GenericObject:
      type: object
      additionalProperties: true